Remove more duplicated code. Group common code into shared library.
