+++ projects/rocprofiler-systems/source/bin/CMakeLists.txt
@@ -4,6 +4,9 @@ set(ROCPROFSYS_EXE_INSTALL_RPATH
     "\$ORIGIN/../${CMAKE_INSTALL_LIBDIR}:\$ORIGIN/../${CMAKE_INSTALL_LIBDIR}/${PROJECT_NAME}"
 )
 
+# common library (shared by run, sample, etc.)
+add_subdirectory(common)
+
 # executables
 add_subdirectory(rocprof-sys-avail)
 add_subdirectory(rocprof-sys-causal)
