+++ projects/rocprofiler-systems/source/bin/rocprof-sys-sample/CMakeLists.txt
@@ -24,6 +24,7 @@ target_link_libraries(
         rocprofiler-systems::rocprofiler-systems-compile-definitions
         rocprofiler-systems::rocprofiler-systems-headers
         rocprofiler-systems::rocprofiler-systems-common-library
+        rocprofiler-systems::rocprofiler-systems-bin-common
         rocprofiler-systems::rocprofiler-systems-json
 )
 set_target_properties(
