+++ projects/rocprofiler-systems/source/tests/CMakeLists.txt
@@ -28,4 +28,5 @@ target_link_libraries(
         rocprofiler-systems-object-library
         rocprofiler-systems-pmc-library
         rocprofiler-systems::rocprofiler-systems-binary
+        $<$<TARGET_EXISTS:rocprofiler-systems-bin-common>:rocprofiler-systems::rocprofiler-systems-bin-common>
 )
